change form table to ssr
